implemeted floating point constructor functions
If a float constant is declared in the C code, the needed functions to
declare it are generated

+++ trunk/examples/programs/FloatingPoint/float.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-  float x = 1.231;
-  float y = 2.623;
+  double x = 1.231;
+  double y = 2.623;
   if (x == y)
   {
     return 1;
